Improved code organization, moved a couple of functions to bokehcamdisplay.py
